+++ CHANGES.rst
@@ -5,8 +5,10 @@ Version 1.2.4
 -------------
 
 * Added support for multiple merchant accounts in a single database (multi-tenant support).
-* Added methods to ``DocdataOrder.objects``: ``current_merchant()``, ``for_reference()``, ``for_order()``.
+* Added manager methods to ``DocdataOrder.objects``: ``current_merchant()``, ``for_reference()``, ``for_order()``.
+* Added support for the ``Invoice`` element. This is needed to pass state for PayPal.
 * Added ``DOCDATA_FACADE_CLASS`` setting, so all views use a custom Facade
+* Added ``DOCDATA_HOUSE_NUMBER_FIELD`` setting.
 * Added ``oscar_docdata.gateway.to_iso639_part1()`` function, for sending language codes in the proper format.
 * Added pagination in docdata dashboard view.
 * Fix detecting expired orders via the status API.